moved storage to utils cleaned up comments updated AGENTS.md
